refactor(learningDiaryOverview): puts some components and functions into the table.tsx
